Tweak avatar physics values to make them behave better
THIS IS A SHARED EXPERIENCE CHANGE.

I am performing this change because the avatar physics implementation
is counter-intuitive at best, and horribly broken at worse:

- Full dampening does not actually fully dampen.
- Physics effect strength seems to depend on client FPS in some cases
- Achieving natural avatar physics with references is impossible

My goal with this refactor is to fix the issues mentioned above, making
avatar physics something useful instead of a comedic eye-candy.

Misc: damping is the wrong word, it should have been dampening.